+++ pyvo/mivot/tests/test_mango_annoter.py
@@ -15,6 +15,7 @@ from pyvo.utils import activate_features
 from pyvo.mivot.version_checker import check_astropy_version
 from pyvo.mivot.utils.xml_utils import XmlUtils
 from pyvo.mivot.writer.instances_from_models import InstancesFromModels
+from pyvo.mivot.utils.exceptions import MappingError
 
 
 # Enable MIVOT-specific features in the pyvo library
@@ -71,6 +72,17 @@ def add_color(builder):
 
 
 @pytest.mark.filterwarnings("ignore:root:::")
+def add_color_without_definition(builder):
+
+    filter_ids = {"high": "GAIA/GAIA3.Grp/AB", "low": "GAIA/GAIA3.Grvs/AB"}
+    mapping = {"value": 8.76,
+             "error": {"class": "PErrorAsym1D", "low": 1, "high": 3}
+             }
+    semantics = {"description": "very nice color", "uri": "vocabulary#term", "label": "term"}
+    builder.add_mango_color(filter_ids=filter_ids, mapping=mapping, semantics=semantics)
+
+
+@pytest.mark.filterwarnings("ignore:root:::")
 def add_photometry(builder):
     photcal_id = "GAIA/GAIA3.Grvs/AB"
     mapping = {"value": "GRVSmag",
@@ -82,13 +94,13 @@ def add_photometry(builder):
     builder.add_mango_brightness(photcal_id=photcal_id, mapping=mapping, semantics=semantics)
 
 
-def add_epoch_positon(builder):
+def add_epoch_position(builder):
     frames = {"spaceSys": {"spaceRefFrame": "ICRS", "refPosition": 'BARYCENTER', "equinox": None},
              "timeSys": {"timescale": "TCB", "refPosition": 'BARYCENTER'}}
     mapping = {"longitude": "_RAJ2000", "latitude": "_DEJ2000",
              "pmLongitude": "pmRA", "pmLatitude": "pmDE",
              "parallax": "Plx", "radialVelocity": "RV",
-             "obsDate": {"representation": "mjd", "dateTime": 579887.6},
+             "obsDate": {"dmtype": "mango:mjd", "value": 579887.6},
              "correlations": {"isCovariance": True,
                               "longitudeLatitude": "RADEcor",
                               "latitudePmLongitude": "DEpmRAcor", "latitudePmLatitude": "DEpmDEcor",
@@ -101,6 +113,19 @@ def add_epoch_positon(builder):
                          "parallax": {"class": "PErrorSym1D", "sigma": "e_Plx"},
                          "radialVelocity": {"class": "PErrorSym1D", "sigma": "e_RV"}
                         }
+             }
+    semantics = {"description": "6 parameters position",
+                 "uri": "https://www.ivoa.net/rdf/uat/2024-06-25/uat.html#astronomical-location",
+                 "label": "Astronomical location"}
+
+    builder.add_mango_epoch_position(frames=frames, mapping=mapping, semantics=semantics)
+
+
+def add_epoch_position_with_wrong_date(builder):
+    frames = {"spaceSys": {"spaceRefFrame": "ICRS", "refPosition": 'BARYCENTER', "equinox": None},
+             "timeSys": {"timescale": "TCB", "refPosition": 'BARYCENTER'}}
+    mapping = {"longitude": "_RAJ2000", "latitude": "_DEJ2000",
+             "obsDate": {"representation": "mango:mjd", "value": 579887.6},
              }
     semantics = {"description": "6 parameters position",
                  "uri": "https://www.ivoa.net/rdf/uat/2024-06-25/uat.html#astronomical-location",
@@ -151,11 +176,18 @@ def test_all_properties():
           })
     add_color(builder)
     add_photometry(builder)
-    add_epoch_positon(builder)
+    add_epoch_position(builder)
     builder.pack_into_votable(schema_check=False)
     assert XmlUtils.strip_xml(builder._annotation.mivot_block) == (
         XmlUtils.strip_xml(get_pkg_data_contents("data/reference/mango_object.xml"))
     )
+
+    builder = InstancesFromModels(votable, dmid="DR3Name")
+    with pytest.raises(MappingError):
+        add_epoch_position_with_wrong_date(builder)
+
+    with pytest.raises(MappingError):
+        add_color_without_definition(builder)
 
 
 @pytest.mark.skipif(not check_astropy_version(), reason="need astropy 6+")